+++ index.js
@@ -83,20 +83,35 @@ var defer = typeof setImmediate === 'function'
  * @public
  */
 
-function session(options){
-  var options = options || {}
-  //  name - previously "options.key"
-    , name = options.name || options.key || 'connect.sid'
-    , store = options.store || new MemoryStore
-    , trustProxy = options.proxy
-    , storeReady = true
-    , rollingSessions = options.rolling || false;
-  var cookieOptions = options.cookie || {};
-  var resaveSession = options.resave;
-  var saveUninitializedSession = options.saveUninitialized;
-  var secret = options.secret;
+function session(options) {
+  var opts = options || {}
 
-  var generateId = options.genid || generateSessionId;
+  // get the cookie options
+  var cookieOptions = opts.cookie || {}
+
+  // get the session id generate function
+  var generateId = opts.genid || generateSessionId
+
+  // get the session cookie name
+  var name = opts.name || opts.key || 'connect.sid'
+
+  // get the session store
+  var store = opts.store || new MemoryStore()
+
+  // get the trust proxy setting
+  var trustProxy = opts.proxy
+
+  // get the resave session option
+  var resaveSession = opts.resave;
+
+  // get the rolling session option
+  var rollingSessions = Boolean(opts.rolling)
+
+  // get the save uninitialized session option
+  var saveUninitializedSession = opts.saveUninitialized
+
+  // get the cookie signing secret
+  var secret = opts.secret
 
   if (typeof generateId !== 'function') {
     throw new TypeError('genid option must be a function');
@@ -112,12 +127,12 @@ function session(options){
     saveUninitializedSession = true;
   }
 
-  if (options.unset && options.unset !== 'destroy' && options.unset !== 'keep') {
+  if (opts.unset && opts.unset !== 'destroy' && opts.unset !== 'keep') {
     throw new TypeError('unset option must be "destroy" or "keep"');
   }
 
   // TODO: switch to "destroy" on next major
-  var unsetDestroy = options.unset === 'destroy';
+  var unsetDestroy = opts.unset === 'destroy'
 
   if (Array.isArray(secret) && secret.length === 0) {
     throw new TypeError('secret option array must contain one or more strings');
@@ -151,7 +166,8 @@ function session(options){
 
   var storeImplementsTouch = typeof store.touch === 'function';
 
-  // register event listeners for the store
+  // register event listeners for the store to track readiness
+  var storeReady = true
   store.on('disconnect', function ondisconnect() {
     storeReady = false
   })